working username and pfp size warnings
